+++ src/test/kotlin/de/qualersoft/jmeter/gradleplugin/task/JMeterRunTaskTest.kt
@@ -1,6 +1,7 @@
 package de.qualersoft.jmeter.gradleplugin.task
 
 import de.qualersoft.jmeter.gradleplugin.entryEndsWith
+import de.qualersoft.jmeter.gradleplugin.matchingEntry
 import io.kotest.assertions.withClue
 import io.kotest.matchers.collections.contain
 import io.kotest.matchers.should
@@ -25,7 +26,7 @@ class JMeterRunTaskTest : JMeterTaskTestBase() {
       { withClue("No gui flag") { args should contain("-n") } },
       { withClue("Enable report flag") { args should contain("-e") } },
       { withClue("Output param") { args should contain("-o") } },
-      { args shouldHave entryEndsWith("reports${File.separatorChar}jmeter") }
+      { args shouldHave entryEndsWith("reports${File.separatorChar}jmeter${File.separatorChar}Test") }
     )
   }
 
@@ -42,5 +43,50 @@ class JMeterRunTaskTest : JMeterTaskTestBase() {
       { withClue("Output param") { args shouldNot contain("-o") } },
       { args shouldNotHave entryEndsWith("reports${File.separatorChar}jmeter") }
     )
+  }
+
+  @Test
+  fun argsWithGlobalPropertyFile() {
+    val propFile = "GlobPropFile.properties"
+    val task = createTaskWithConfig<JMeterRunTask>({}, {
+      globalPropertiesFile.set(File(propFile))
+      jmxFile.set("Report.jmx")
+    }).get()
+
+    val args = task.createRunArguments()
+    args shouldHave matchingEntry("-G[^=]+$propFile".toRegex())
+  }
+
+  @Test
+  fun argsWithGlobalProperties() {
+    val task = createTaskWithConfig<JMeterRunTask>({}, {
+      globalProperties.put("Global", "property")
+      jmxFile.set("Report.jmx")
+    }).get()
+
+    val args = task.createRunArguments()
+    args should contain("-GGlobal=property")
+  }
+
+  @Test
+  fun withDeleteFlag() {
+    val task = createTaskWithConfig<JMeterRunTask>({}, {
+      deleteResults = true
+      jmxFile.set("Report.jmx")
+    }).get()
+
+    val args = task.createRunArguments()
+    args should contain("-f")
+  }
+
+  @Test
+  fun withoutDeleteFlag() {
+    val task = createTaskWithConfig<JMeterRunTask>({}, {
+      deleteResults = false
+      jmxFile.set("Report.jmx")
+    }).get()
+
+    val args = task.createRunArguments()
+    args shouldNot contain("-f")
   }
 }